# docs/api/openapi_v1.json
{
  "components": {
    "schemas": {
...
      "CashFlowCreate": {
        "properties": {
          "amount": {
            "description": "Derived cashflow amount",
            "title": "Amount",
            "type": "number"
          },
          "contract_id": {
            "description": "Source contract ID",
            "format": "uuid",
            "title": "Contract Id",
            "type": "string"
          },
          "value_date": {
            "description": "Cashflow value date",
            "format": "date",
            "title": "Value Date",
            "type": "string"
          },
          "view": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CashFlowView"
              }
            ],
            "description": "CashFlow view type"
          }
        },
        "required": [
          "contract_id",
          "view",
          "amount",
          "value_date"
        ],
        "title": "CashFlowCreate",
        "type": "object"
      },
...
      "CashFlowRead": {
        "properties": {
          "amount": {
            "description": "Derived cashflow amount",
            "title": "Amount",
            "type": "number"
          },
          "contract_id": {
            "description": "Source contract ID",
            "format": "uuid",
            "title": "Contract Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "value_date": {
            "description": "Cashflow value date",
            "format": "date",
            "title": "Value Date",
            "type": "string"
          },
          "view": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CashFlowView"
              }
            ],
            "description": "CashFlow view type"
          }
        },
        "required": [
          "contract_id",
          "view",
          "amount",
          "value_date",
          "id",
          "created_at"
        ],
        "title": "CashFlowRead",
        "type": "object"
      },
      "CashFlowView": {
        "enum": [
          "analytic",
          "baseline",
          "ledger",
          "what_if"
        ],
        "title": "CashFlowView",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HedgeOrderLinkageCreate": {
        "properties": {
          "contract_id": {
            "description": "Linked hedge contract ID",
            "format": "uuid",
            "title": "Contract Id",
            "type": "string"
          },
          "order_id": {
            "description": "Linked order ID",
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "quantity_mt": {
            "description": "Linked quantity in MT",
            "title": "Quantity Mt",
            "type": "number"
          }
        },
        "required": [
          "order_id",
          "contract_id",
          "quantity_mt"
        ],
        "title": "HedgeOrderLinkageCreate",
        "type": "object"
      },
      "HedgeOrderLinkageRead": {
        "properties": {
          "contract_id": {
            "format": "uuid",
            "title": "Contract Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "order_id": {
            "format": "uuid",
            "title": "Order Id",
            "type": "string"
          },
          "quantity_mt": {
            "title": "Quantity Mt",
            "type": "number"
          }
        },
        "required": [
          "id",
          "order_id",
          "contract_id",
          "quantity_mt",
          "created_at"
        ],
        "title": "HedgeOrderLinkageRead",
        "type": "object"
      },
...
      "RFQCreate": {
        "properties": {
          "buy_trade_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Referenced buy trade (RFQ id) for SPREAD",
            "title": "Buy Trade Id"
          },
          "commodity": {
            "title": "Commodity",
            "type": "string"
          },
          "delivery_window_end": {
            "format": "date",
            "title": "Delivery Window End",
            "type": "string"
          },
          "delivery_window_start": {
            "format": "date",
            "title": "Delivery Window Start",
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/RFQDirection"
          },
          "intent": {
            "$ref": "#/components/schemas/RFQIntent"
          },
          "invitations": {
            "items": {
              "$ref": "#/components/schemas/RFQInvitationCreate"
            },
            "title": "Invitations",
            "type": "array"
          },
          "order_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Referenced commercial order ID",
            "title": "Order Id"
          },
          "quantity_mt": {
            "description": "Quantity in metric tons (MT)",
            "title": "Quantity Mt",
            "type": "number"
          },
          "sell_trade_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Referenced sell trade (RFQ id) for SPREAD",
            "title": "Sell Trade Id"
          }
        },
        "required": [
          "intent",
          "commodity",
          "quantity_mt",
          "delivery_window_start",
          "delivery_window_end",
          "direction"
        ],
        "title": "RFQCreate",
        "type": "object"
      },
      "RFQDirection": {
        "enum": [
          "BUY",
          "SELL"
        ],
        "title": "RFQDirection",
        "type": "string"
      },
      "RFQIntent": {
        "enum": [
          "COMMERCIAL_HEDGE",
          "GLOBAL_POSITION",
          "SPREAD"
        ],
        "title": "RFQIntent",
        "type": "string"
      },
      "RFQInvitationChannel": {
        "enum": [
          "email",
          "api",
          "whatsapp",
          "bank",
          "broker",
          "other"
        ],
        "title": "RFQInvitationChannel",
        "type": "string"
      },
      "RFQInvitationCreate": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/RFQInvitationChannel"
          },
          "idempotency_key": {
            "description": "Idempotency key for send",
            "title": "Idempotency Key",
            "type": "string"
          },
          "message_body": {
            "description": "Exact message body sent",
            "title": "Message Body",
            "type": "string"
          },
          "provider_message_id": {
            "description": "Provider message id",
            "title": "Provider Message Id",
            "type": "string"
          },
          "recipient_id": {
            "description": "Recipient identifier",
            "title": "Recipient Id",
            "type": "string"
          },
          "recipient_name": {
            "description": "Recipient name",
            "title": "Recipient Name",
            "type": "string"
          },
          "send_status": {
            "$ref": "#/components/schemas/RFQInvitationStatus"
          },
          "sent_at": {
            "format": "date-time",
            "title": "Sent At",
            "type": "string"
          }
        },
        "required": [
          "recipient_id",
          "recipient_name",
          "channel",
          "message_body",
          "provider_message_id",
          "send_status",
          "sent_at",
          "idempotency_key"
        ],
        "title": "RFQInvitationCreate",
        "type": "object"
      },
      "RFQInvitationRead": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/RFQInvitationChannel"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "idempotency_key": {
            "description": "Idempotency key for send",
            "title": "Idempotency Key",
            "type": "string"
          },
          "message_body": {
            "description": "Exact message body sent",
            "title": "Message Body",
            "type": "string"
          },
          "provider_message_id": {
            "description": "Provider message id",
            "title": "Provider Message Id",
            "type": "string"
          },
          "recipient_id": {
            "description": "Recipient identifier",
            "title": "Recipient Id",
            "type": "string"
          },
          "recipient_name": {
            "description": "Recipient name",
            "title": "Recipient Name",
            "type": "string"
          },
          "rfq_id": {
            "format": "uuid",
            "title": "Rfq Id",
            "type": "string"
          },
          "rfq_number": {
            "title": "Rfq Number",
            "type": "string"
          },
          "send_status": {
            "$ref": "#/components/schemas/RFQInvitationStatus"
          },
          "sent_at": {
            "format": "date-time",
            "title": "Sent At",
            "type": "string"
          }
        },
        "required": [
          "recipient_id",
          "recipient_name",
          "channel",
          "message_body",
          "provider_message_id",
          "send_status",
          "sent_at",
          "idempotency_key",
          "id",
          "rfq_id",
          "rfq_number",
          "created_at"
        ],
        "title": "RFQInvitationRead",
        "type": "object"
      },
      "RFQInvitationStatus": {
        "enum": [
          "queued",
          "sent",
          "failed"
        ],
        "title": "RFQInvitationStatus",
        "type": "string"
      },
...
      "RFQRead": {
        "properties": {
          "buy_trade_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buy Trade Id"
          },
          "commercial_active_mt": {
            "title": "Commercial Active Mt",
            "type": "number"
          },
          "commercial_net_mt": {
            "title": "Commercial Net Mt",
            "type": "number"
          },
          "commercial_passive_mt": {
            "title": "Commercial Passive Mt",
            "type": "number"
          },
          "commercial_reduction_applied_mt": {
            "title": "Commercial Reduction Applied Mt",
            "type": "number"
          },
          "commodity": {
            "title": "Commodity",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivery_window_end": {
            "format": "date",
            "title": "Delivery Window End",
            "type": "string"
          },
          "delivery_window_start": {
            "format": "date",
            "title": "Delivery Window Start",
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/RFQDirection"
          },
          "exposure_snapshot_timestamp": {
            "format": "date-time",
            "title": "Exposure Snapshot Timestamp",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "intent": {
            "$ref": "#/components/schemas/RFQIntent"
          },
          "invitations": {
            "items": {
              "$ref": "#/components/schemas/RFQInvitationRead"
            },
            "title": "Invitations",
            "type": "array"
          },
          "order_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "quantity_mt": {
            "title": "Quantity Mt",
            "type": "number"
          },
          "rfq_number": {
            "title": "Rfq Number",
            "type": "string"
          },
          "sell_trade_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sell Trade Id"
          },
          "state": {
            "$ref": "#/components/schemas/RFQState"
          }
        },
        "required": [
          "id",
          "rfq_number",
          "intent",
          "commodity",
          "quantity_mt",
          "delivery_window_start",
          "delivery_window_end",
          "direction",
          "order_id",
          "buy_trade_id",
          "sell_trade_id",
          "commercial_active_mt",
          "commercial_passive_mt",
          "commercial_net_mt",
          "commercial_reduction_applied_mt",
          "exposure_snapshot_timestamp",
          "state",
          "created_at"
        ],
        "title": "RFQRead",
        "type": "object"
      },
...
      "RFQState": {
        "enum": [
          "CREATED",
          "SENT",
          "QUOTED",
          "AWARDED",
          "CLOSED"
        ],
        "title": "RFQState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Hedge Control Platform",
    "version": "0.3.0-phase3-step2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cashflows": {
      "get": {
        "operationId": "list_cashflows_cashflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CashFlowRead"
                  },
                  "title": "Response List Cashflows Cashflows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Cashflows",
        "tags": [
          "CashFlows"
        ]
      },
      "post": {
        "operationId": "create_cashflow_cashflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashFlowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Cashflow",
        "tags": [
          "CashFlows"
        ]
      }
    },
...
    "/linkages": {
      "post": {
        "operationId": "create_linkage_linkages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HedgeOrderLinkageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HedgeOrderLinkageRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Linkage",
        "tags": [
          "Linkages"
        ]
      }
    },
...
    "/rfqs": {
      "post": {
        "operationId": "create_rfq_rfqs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RFQCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RFQRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rfq",
        "tags": [
          "RFQs"
        ]
      }
    },
...
  }
}